Исправил недочёты
